+++ cmake/GenerateGrpc.cmake
@@ -2,7 +2,11 @@
 #--------------------
 
 # Set the path to the directory where you want to generate the C++ headers
-set(GENERATED_DIR ${CMAKE_CURRENT_SOURCE_DIR})
+# Generated files go in the build directory, not the source directory
+set(GENERATED_DIR ${CMAKE_BINARY_DIR}/src/generated)
+
+# Create the generated directory if it doesn't exist
+file(MAKE_DIRECTORY ${GENERATED_DIR})
 
 # Set the path to the directory containing the .proto files
 set(PROTO_DIR ${CMAKE_SOURCE_DIR}/proto)
@@ -18,11 +22,11 @@ set(PROTO_FILES_DEPEND "")
 foreach(elem ${PROTO_FILES})
     string(REPLACE ".proto" "" file_no_ext ${elem})
     list(APPEND PROTO_FILES_DEPEND ${CMAKE_SOURCE_DIR}/proto/${elem})
-    list(APPEND PROTO_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/${file_no_ext}.pb.h)
-    list(APPEND GRPC_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/${file_no_ext}.grpc.pb.h)
-    list(APPEND GRPC_MOCK_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/${file_no_ext}_mock.grpc.pb.h)
-    list(APPEND PROTO_SRC ${CMAKE_CURRENT_SOURCE_DIR}/${file_no_ext}.pb.cc)
-    list(APPEND GRPC_SRC ${CMAKE_CURRENT_SOURCE_DIR}/${file_no_ext}.grpc.pb.cc)
+    list(APPEND PROTO_HEADERS ${GENERATED_DIR}/${file_no_ext}.pb.h)
+    list(APPEND GRPC_HEADERS ${GENERATED_DIR}/${file_no_ext}.grpc.pb.h)
+    list(APPEND GRPC_MOCK_HEADERS ${GENERATED_DIR}/${file_no_ext}_mock.grpc.pb.h)
+    list(APPEND PROTO_SRC ${GENERATED_DIR}/${file_no_ext}.pb.cc)
+    list(APPEND GRPC_SRC ${GENERATED_DIR}/${file_no_ext}.grpc.pb.cc)
 endforeach()
 
 # Add a custom command to generate the C++ headers and sources from the .proto files
@@ -33,7 +37,7 @@ add_custom_command(
         --proto_path=${PROTO_DIR}
         --grpc_out=generate_mock_code=true:${GENERATED_DIR}
         --cpp_out=${GENERATED_DIR}
-        --plugin=protoc-gen-grpc=`which grpc_cpp_plugin`
+        --plugin=protoc-gen-grpc=${GRPC_CPP_PLUGIN}
         ${PROTO_FILES}
     DEPENDS ${PROTO_FILES_DEPEND}
     COMMENT "Generating C++ stubs from protobuf files"
@@ -46,10 +50,10 @@ add_custom_target(GenerateGrpcFiles
 )
 
 install(FILES ${PROTO_FILES_DEPEND}
-    DESTINATION proto
+    DESTINATION ${CMAKE_INSTALL_DATADIR}/philote/proto
 )
 install(FILES ${PROTO_HEADERS} ${GRPC_HEADERS} ${GRPC_MOCK_HEADERS}
-        DESTINATION include
+        DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/philote
 )
 
 # Create the GrpcGenerated library using the generated sources
@@ -58,4 +62,10 @@ add_library(GrpcGenerated OBJECT ${PROTO_SRC} ${GRPC_SRC})
 # Ensure the library depends on the generation target
 add_dependencies(GrpcGenerated GenerateGrpcFiles)
 
+target_include_directories(GrpcGenerated
+    PRIVATE
+        ${CMAKE_SOURCE_DIR}/include
+        ${GENERATED_DIR}
+)
 target_link_libraries(GrpcGenerated PRIVATE protobuf::libprotobuf gRPC::grpc++)
+enable_coverage(GrpcGenerated)